requirements.txt
# Remove the conflict markers and keep your requirements
sentence-transformers>=2.2.2
faiss-cpu>=1.7.4
llama-cpp-python>=0.2.11
huggingface-hub>=0.16.4
numpy>=1.21.0
streamlit>=1.28.0
PyMuPDF>=1.23.0
beautifulsoup4>=4.12.0
nbformat>=5.7.0
GitPython>=3.1.30
PyYAML>=6.0
requests>=2.28.0
tqdm>=4.64.0